usb: cdns3: starfive: Simplify mode init
The syscon regmap and offset to the USB mode register is only used at
probe time, so there is no need to store it in the device data. Just get
the regmap pointer in the cdns_mode_init() function where it is needed.
Also this function never uses the platform device, so just pass the
device pointer directly.

Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>